+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New Debugger class and ability to trigger Debugger output from module config screen (superusers only).
+
 ## [0.21.0] - 2020-05-11
 
 ### Added